pyproject.toml
[tool.poetry]
name = "repl_python3_Lichess-Searcher"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-lichess = "^0.10"
python-chess = "^1.999"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
